WIP: Identified critical float/double precision issue in inference and training pipeline
CRITICAL FINDINGS:
- Model structures correctly use double* for weights
- BUT all layer operations (attention, feedforward, layernorm) use float*
- This causes constant double<->float conversions
- Results in precision loss and potential numerical instability
- Explains NaN issues and poor training performance

CHANGES IN THIS COMMIT:
- Fixed memory management in cllm_free()
- Fixed NaN detection in inference
- Added comprehensive debugging tools
- Documented the float/double issue

NEXT STEPS:
- Convert ALL layer operations to use double precision
- Remove float* from all inference/training functions
- Update SIMD operations to support double
- Rebuild and test entire pipeline

+++ include/cllm_inference.h
@@ -47,7 +47,7 @@ CLLMInference* cllm_inference_init(CLLMModel* model);
 int cllm_tokenize(CLLMInference* inference, const char* text, uint32_t* tokens, int max_tokens);
 void cllm_detokenize(CLLMInference* inference, uint32_t* tokens, int num_tokens, char* output, int max_length);
 void cllm_get_embedding(CLLMInference* inference, uint32_t token_id, float* embedding);
-void cllm_apply_positional_encoding(CLLMInference* inference, float* embedding, int position);
+void cllm_apply_positional_encoding(CLLMInference* inference, double* embedding, int position);
 
 /* Embedding layer functions */
 void cllm_embed_token(CLLMInference* inf, uint32_t token_id, float* output);
